feat(ips.byoip): add new route, create new hook for header
ref: #MANAGER-17441
