+++ packages/manager/apps/ips/public/translations/byoip/Messages_fr_FR.json
@@ -1,0 +1,4 @@
+{
+  "title": "Bring Your Own IP (BYOIP)",
+  "description": "Importez vos propres adresses IP publiques et utilisez les sur les services OVHcloud éligibles."
+}
