+++ packages/manager/apps/ips/public/translations/common/Messages_fr_FR.json
@@ -1,0 +1,4 @@
+{
+  "back_link": "Retour à la page précédente",
+  "ips_dashboard_guide": "Tous les guides"
+}
