+++ packages/manager/apps/ips/public/translations/ips/Messages_fr_FR.json
@@ -1,5 +1,6 @@
 {
   "breadcrumb_root_label": "Toutes les IPs",
   "order": "Commander une IP",
+  "byoip": "BYOIP",
   "free_price": "Gratuit"
 }
