+++ packages/manager/apps/ips/public/translations/listing/Messages_fr_FR.json
@@ -1,6 +1,5 @@
 {
   "title": "Gérer les IPs",
-  "ips_dashboard_guide": "Tous les guides",
   "listingTabIpTitle": "Toutes les IP",
   "listingTabManageIpOrganisationsTitle": "Gérer mes organisations",
   "listingTabIpDescription": "Mes adresses IP publiques et services associés",
