+++ packages/manager/apps/ips/public/translations/order/Messages_fr_FR.json
@@ -1,6 +1,5 @@
 {
   "title": "Commander des Additional IP",
-  "back_link": "Retour à la page précédente",
   "per_ip": "/IP",
   "per_ip_full": "par adresse IP",
   "ip_version_title": "Sélectionner la version de l'adresse IP",
